Create consoleresize.c
This has nothing to do with the CD-berry, but might come in handy when accessing a Raspberry Pi using the serial console. After compiling the resulting binary resizes the rows and colums of the serial console to the current window size. To see the result, resize your Putty console, start nano, exit nano, run consoleresize, then run nano again